+++ apps/crisis-unleashed/app/admin/logo-export/page.tsx
@@ -6,7 +6,9 @@ import { Button } from "@/components/ui/button"
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select"
 import { Label } from "@/components/ui/label"
 import { Input } from "@/components/ui/input"
-import { LogoSystem, type LogoVariant, type LogoSize } from "@/components/logo-system"
+// Import directly from component files instead of barrel file
+import LogoSystem from "@/components/logo-system/logo-system"
+import type { LogoVariant, LogoSize } from "@/components/logo-system/logo-variant"
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs"
 import { Switch } from "@/components/ui/switch"
 import { Download, Copy, Check } from "lucide-react"
