+++ apps/crisis-unleashed/app/admin/logo-variants/page.tsx
@@ -1,7 +1,9 @@
 "use client"
 
 import { useState } from "react"
-import { LogoSystem, type LogoVariant, type LogoSize } from "@/components/logo-system"
+// Import directly from component files instead of barrel file
+import LogoSystem from "@/components/logo-system/logo-system"
+import type { LogoVariant, LogoSize } from "@/components/logo-system/logo-variant"
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs"
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select"
@@ -9,6 +11,7 @@ import { Switch } from "@/components/ui/switch"
 import { Label } from "@/components/ui/label"
 import { Button } from "@/components/ui/button"
 import { useTheme } from "@/contexts/theme-context"
+import LogoUsageGuide from "@/components/logo-system/logo-usage-guide"
 
 const factions = [
   "cybernetic-nexus",
@@ -203,108 +206,5 @@ export default function LogoVariantsPage() {
         </TabsContent>
       </Tabs>
     </div>
-  )
-}
-
-/**
- * Displays detailed guidelines and visual examples for each logo variant.
- *
- * Provides usage recommendations and large previews for standard, compact, horizontal, vertical, icon-only, text-only, footer, mobile, print, and watermark logo variants to guide appropriate application in different contexts.
- */
-function LogoUsageGuide() {
-  return (
-    <Card>
-      <CardHeader>
-        <CardTitle>Logo Usage Guidelines</CardTitle>
-      </CardHeader>
-      <CardContent className="space-y-6">
-        <div>
-          <h3 className="text-lg font-semibold mb-2">Standard Logo</h3>
-          <p>
-            Use the standard logo for most applications where space permits. This is the primary logo representation.
-          </p>
-          <div className="mt-4 p-4 bg-muted/20 rounded-md flex justify-center">
-            <LogoSystem variant="standard" size="lg" />
-          </div>
-        </div>
-
-        <div>
-          <h3 className="text-lg font-semibold mb-2">Compact Logo</h3>
-          <p>
-            Use the compact logo in spaces where the standard logo would be too large but you still need both icon and
-            text.
-          </p>
-          <div className="mt-4 p-4 bg-muted/20 rounded-md flex justify-center">
-            <LogoSystem variant="compact" size="lg" />
-          </div>
-        </div>
-
-        <div>
-          <h3 className="text-lg font-semibold mb-2">Horizontal Logo</h3>
-          <p>Use the horizontal logo in headers and navigation bars where horizontal space is available.</p>
-          <div className="mt-4 p-4 bg-muted/20 rounded-md flex justify-center">
-            <LogoSystem variant="horizontal" size="lg" />
-          </div>
-        </div>
-
-        <div>
-          <h3 className="text-lg font-semibold mb-2">Vertical Logo</h3>
-          <p>Use the vertical logo in narrow spaces or for decorative purposes where vertical space is available.</p>
-          <div className="mt-4 p-4 bg-muted/20 rounded-md flex justify-center">
-            <LogoSystem variant="vertical" size="lg" />
-          </div>
-        </div>
-
-        <div>
-          <h3 className="text-lg font-semibold mb-2">Icon-Only Logo</h3>
-          <p>Use the icon-only logo for app icons, favicons, and small UI elements where text would be illegible.</p>
-          <div className="mt-4 p-4 bg-muted/20 rounded-md flex justify-center">
-            <LogoSystem variant="icon-only" size="lg" />
-          </div>
-        </div>
-
-        <div>
-          <h3 className="text-lg font-semibold mb-2">Text-Only Logo</h3>
-          <p>
-            Use the text-only logo when the icon is displayed separately or when you need to emphasize the brand name.
-          </p>
-          <div className="mt-4 p-4 bg-muted/20 rounded-md flex justify-center">
-            <LogoSystem variant="text-only" size="lg" />
-          </div>
-        </div>
-
-        <div>
-          <h3 className="text-lg font-semibold mb-2">Footer Logo</h3>
-          <p>Use the footer logo in website footers, typically in a more subtle presentation.</p>
-          <div className="mt-4 p-4 bg-muted/20 rounded-md flex justify-center">
-            <LogoSystem variant="footer" size="lg" />
-          </div>
-        </div>
-
-        <div>
-          <h3 className="text-lg font-semibold mb-2">Mobile Logo</h3>
-          <p>Use the mobile logo for mobile navigation and small screen contexts.</p>
-          <div className="mt-4 p-4 bg-muted/20 rounded-md flex justify-center">
-            <LogoSystem variant="mobile" size="lg" />
-          </div>
-        </div>
-
-        <div>
-          <h3 className="text-lg font-semibold mb-2">Print Logo</h3>
-          <p>Use the print logo for printed materials, ensuring high contrast and readability.</p>
-          <div className="mt-4 p-4 bg-muted/20 rounded-md flex justify-center">
-            <LogoSystem variant="print" size="lg" monochrome />
-          </div>
-        </div>
-
-        <div>
-          <h3 className="text-lg font-semibold mb-2">Watermark Logo</h3>
-          <p>Use the watermark logo as a subtle background element or for document watermarking.</p>
-          <div className="mt-4 p-4 bg-muted/20 rounded-md flex justify-center">
-            <LogoSystem variant="watermark" size="lg" />
-          </div>
-        </div>
-      </CardContent>
-    </Card>
   )
 }
