+++ apps/crisis-unleashed/components/logo-system/logo-system.tsx
@@ -1,41 +1,23 @@
 "use client"
 
-import { ResponsiveAnimatedLogo } from "@/components/responsive-animated-logo"
-import { useMemo } from "react"
-import type { LogoSize, LogoVariant } from "./logo-variant"
-
-export interface LogoSystemProps {
-  variant?: LogoVariant
-  size?: LogoSize
-  faction?: string
-  animated?: boolean
-  interactive?: boolean
-  monochrome?: boolean
-  inverted?: boolean
-  withTagline?: boolean
-  className?: string
-}
-
-// Map LogoVariant to ResponsiveAnimatedLogo variant
-const variantMap = {
-  standard:  "standard",
-  "icon-only": "icon-only",
-  horizontal: "horizontal",
-  vertical:   "vertical",
-  wordmark:   "text-only",
-  badge:      "icon-only",
-  minimal:    "compact",
-} as const
-
-type ResponsiveVariant = typeof variantMap[keyof typeof variantMap]
-
-const mapVariantToResponsiveVariant = (v: LogoVariant): ResponsiveVariant =>
-  variantMap[v] ?? "standard";
+import { cn } from "@/lib/utils"
+import { FactionId } from "./utils"
+import LogoVariant, { LogoSize, LogoVariantProps } from "./logo-variant"
 
 /**
- * Renders a responsive, animated logo with configurable appearance and behavior.
- *
- * Maps the provided logo variant to a supported variant for the underlying {@link ResponsiveAnimatedLogo} component, ensuring consistent rendering across different variant options.
+ * LogoSystem is a wrapper around LogoVariant for backward compatibility
+ * It supports all the same props as LogoVariant
+ */
+export interface LogoSystemProps extends Omit<LogoVariantProps, 'variant'> {
+  /**
+   * The variant of the logo to display, including legacy variants
+   */
+  variant?: LogoVariantProps["variant"] | "animated" | "compact" | "icon-only" | "text-only" | "wordmark" | "footer" | "mobile" | "print" | "watermark";
+}
+
+/**
+ * Legacy LogoSystem component that maps to the new LogoVariant component
+ * This ensures backward compatibility with existing code
  */
 export default function LogoSystem({
   variant = "standard",
@@ -47,25 +29,52 @@ export default function LogoSystem({
   inverted = false,
   withTagline = false,
   className,
+  ...props
 }: LogoSystemProps) {
+  // Map legacy variants to new variants
+  let mappedVariant: LogoVariantProps["variant"] = "standard";
+  let isAnimated = animated;
   
-const responsiveVariant = useMemo(
-  () => mapVariantToResponsiveVariant(variant),
-  [variant],
-)
-  
+  // Map the legacy variants to the new system
+  switch (variant) {
+    case "icon-only":
+      mappedVariant = "icon";
+      break;
+    case "text-only":
+    case "wordmark":
+      mappedVariant = "minimal";
+      break;
+    case "compact":
+      mappedVariant = "horizontal";
+      break;
+    case "animated":
+      mappedVariant = "standard";
+      isAnimated = true;
+      break;
+    case "footer":
+    case "mobile":
+    case "print":
+    case "watermark":
+      // These legacy variants can all map to minimal for now
+      mappedVariant = "minimal";
+      break;
+    default:
+      // For standard, horizontal, vertical, icon, and minimal, use as is
+      mappedVariant = variant as LogoVariantProps["variant"];
+  }
+
   return (
-    <ResponsiveAnimatedLogo
-      variant={responsiveVariant}
+    <LogoVariant
+      variant={mappedVariant}
       size={size}
       faction={faction}
+      animated={isAnimated}
       interactive={interactive}
       monochrome={monochrome}
       inverted={inverted}
       withTagline={withTagline}
       className={className}
-      animated={animated}
-      useSvg={true}
+      {...props}
     />
   )
 }
