+++ apps/crisis-unleashed/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "my-v0-project",
+  "name": "vv-crisis-unleasheded",
   "version": "0.1.0",
   "private": true,
   "scripts": {
